make Transformer a single class
