+++ transform/vision/src/main/scala/com/intel/analytics/zoo/transform/vision/image/Types.scala
@@ -178,92 +178,6 @@ object ImageFeature {
   val logger = Logger.getLogger(getClass)
 }
 
-abstract class FeatureTransformer() extends Transformer[ImageFeature, ImageFeature] {
-
-  import FeatureTransformer.logger
-
-  private var outKey: Option[String] = None
-
-  def setOutKey(key: String): this.type = {
-    outKey = Some(key)
-    this
-  }
-
-  protected def transformMat(feature: ImageFeature): Unit = {}
-
-  def transform(feature: ImageFeature): ImageFeature = {
-    if (!feature.isValid) return feature
-    try {
-      transformMat(feature)
-      if (outKey.isDefined) {
-        require(outKey.get != ImageFeature.mat, s"the output key should not equal to" +
-          s" ${ImageFeature.mat}, please give another name")
-        if (feature.contains(outKey.get)) {
-          val mat = feature(outKey.get).asInstanceOf[OpenCVMat]
-          feature.opencvMat().copyTo(mat)
-        } else {
-          feature(outKey.get) = feature.opencvMat().clone()
-        }
-      }
-    } catch {
-      case e: Exception =>
-        val uri = if (feature.contains(ImageFeature.uri)) feature(ImageFeature.uri) else ""
-        logger.warn(s"failed ${uri} in transformer ${getClass}")
-        e.printStackTrace()
-        feature.isValid = false
-    }
-    feature
-  }
-
-  override def apply(prev: Iterator[ImageFeature]): Iterator[ImageFeature] = {
-    prev.map(transform)
-  }
-
-  // scalastyle:off methodName
-  // scalastyle:off noSpaceBeforeLeftBracket
-  def ->(other: FeatureTransformer): FeatureTransformer = {
-    new ChainedFeatureTransformer(this, other)
-  }
-
-  // scalastyle:off methodName
-  // scalastyle:off noSpaceBeforeLeftBracket
-  override def ->[C](other: Transformer[ImageFeature, C]): Transformer[ImageFeature, C] = {
-    new ChainedTransformer(this, other)
-  }
-}
-
-object FeatureTransformer {
-  val logger = Logger.getLogger(getClass)
-}
-
-class ChainedFeatureTransformer(first: FeatureTransformer, last: FeatureTransformer) extends
-  FeatureTransformer {
-
-  override def transform(prev: ImageFeature): ImageFeature = {
-    last.transform(first.transform(prev))
-  }
-}
-
-
-class RandomTransformer(transformer: FeatureTransformer, maxProb: Double)
-  extends FeatureTransformer {
-  override def transform(prev: ImageFeature): ImageFeature = {
-    if (RNG.uniform(0, 1) < maxProb) {
-      transformer.transform(prev)
-    }
-    prev
-  }
-
-  override def toString: String = {
-    s"Random[${transformer.getClass.getCanonicalName}, $maxProb]"
-  }
-}
-
-object RandomTransformer {
-  def apply(transformer: FeatureTransformer, maxProb: Double): RandomTransformer =
-    new RandomTransformer(transformer, maxProb)
-}
-
 object Image {
   def read(path: String, sc: SparkContext): DistributedImageFrame = {
     val images = sc.binaryFiles(path).map { case (p, stream) =>
